Extract fundingState reducer parts from funding reducer
