+++ packages/wallet/src/redux/reducers/channels/funding/index.ts
@@ -1,0 +1,336 @@
+import * as states from '../../../states/channels';
+import * as fundingStates from '../../../states/channels/funding/directFunding';
+import * as actions from '../../../actions';
+import {
+  messageRelayRequested,
+  fundingSuccess,
+  fundingFailure,
+  showWallet,
+  hideWallet,
+  commitmentRelayRequested,
+} from 'magmo-wallet-client/lib/wallet-events';
+
+import { unreachable, validTransition } from '../../../../utils/reducer-utils';
+import { createDepositTransaction } from '../../../../utils/transaction-generator';
+import { signCommitment, validCommitmentSignature } from '../../../../utils/signing-utils';
+
+import { Channel, Commitment, CommitmentType } from 'fmg-core';
+import { handleSignatureAndValidationMessages } from '../../../../utils/state-utils';
+import { bigNumberify } from 'ethers/utils';
+import { NextChannelState } from '../../../states/shared';
+import { directFundingStateReducer } from 'src/redux/reducers/channels/funding/directFundingState';
+
+export const fundingReducer = (
+  state: states.FundingChannelState,
+  action: actions.WalletAction,
+  unhandledAction?: actions.WalletAction,
+): NextChannelState<states.ChannelState> => {
+  // Handle any signature/validation request centrally to avoid duplicating code for each state
+  if (
+    action.type === actions.OWN_COMMITMENT_RECEIVED ||
+    action.type === actions.OPPONENT_COMMITMENT_RECEIVED
+  ) {
+    return {
+      channelState: state,
+      outboxState: { messageOutbox: handleSignatureAndValidationMessages(state, action) },
+    };
+  }
+
+  // TODO: Handle side effects
+  const { fundingState } = directFundingStateReducer(state.fundingState, action, state.channelId);
+
+  state.fundingState = fundingState;
+
+  switch (state.type) {
+    // Setup funding process
+    case states.WAIT_FOR_FUNDING_REQUEST:
+      return waitForFundingRequestReducer(state, action);
+    case states.APPROVE_FUNDING:
+      return approveFundingReducer(state, action, unhandledAction);
+
+    // Funding is ongoing
+    case states.WAIT_FOR_FUNDING_AND_POST_FUND_SETUP:
+      // TODO: Need to define waitForFundingAndPostFundSetup reducer
+      return { channelState: state };
+    case states.WAIT_FOR_FUNDING_CONFIRMATION:
+      return waitForFundingConfirmationReducer(state, action);
+    case states.A_WAIT_FOR_POST_FUND_SETUP:
+      return aWaitForPostFundSetupReducer(state, action);
+    case states.B_WAIT_FOR_POST_FUND_SETUP:
+      return bWaitForPostFundSetupReducer(state, action);
+
+    // Ending the stage
+    case states.ACKNOWLEDGE_FUNDING_SUCCESS:
+      return acknowledgeFundingSuccessReducer(state, action);
+    case states.SEND_FUNDING_DECLINED_MESSAGE:
+      return sendFundingDeclinedMessageReducer(state, action);
+    case states.ACKNOWLEDGE_FUNDING_DECLINED:
+      return acknowledgeFundingDeclinedReducer(state, action);
+    //
+    default:
+      return unreachable(state);
+  }
+};
+
+const acknowledgeFundingDeclinedReducer = (
+  state: states.AcknowledgeFundingDeclined,
+  action: actions.WalletAction,
+): NextChannelState<states.ChannelState> => {
+  switch (action.type) {
+    case actions.FUNDING_DECLINED_ACKNOWLEDGED:
+      return {
+        channelState: states.waitForChannel({
+          ...state,
+        }),
+        outboxState: {
+          messageOutbox: fundingFailure(state.channelId, 'FundingDeclined'),
+          displayOutbox: hideWallet(),
+        },
+      };
+  }
+  return { channelState: state };
+};
+
+const sendFundingDeclinedMessageReducer = (
+  state: states.SendFundingDeclinedMessage,
+  action: actions.WalletAction,
+): NextChannelState<states.WaitForChannel | states.SendFundingDeclinedMessage> => {
+  switch (action.type) {
+    case actions.MESSAGE_SENT:
+      return {
+        channelState: states.waitForChannel({
+          ...state,
+        }),
+        outboxState: {
+          messageOutbox: fundingFailure(state.channelId, 'FundingDeclined'),
+          displayOutbox: hideWallet(),
+        },
+      };
+  }
+  return { channelState: state };
+};
+
+const waitForFundingRequestReducer = (
+  state: states.WaitForFundingRequest,
+  action: actions.WalletAction,
+): NextChannelState<states.ChannelState> => {
+  switch (action.type) {
+    case actions.FUNDING_REQUESTED:
+      return {
+        channelState: states.approveFunding({ ...state }),
+        outboxState: { displayOutbox: showWallet() },
+      };
+    default:
+      return { channelState: state };
+  }
+};
+
+const approveFundingReducer = (
+  state: states.ApproveFunding,
+  action: actions.WalletAction,
+  unhandledAction: actions.WalletAction | undefined,
+): NextChannelState<states.ChannelState> => {
+  switch (action.type) {
+    case actions.FUNDING_RECEIVED_EVENT:
+      return { channelState: states.approveFunding({ ...state }), unhandledAction: action };
+    case actions.FUNDING_APPROVED:
+      if (state.ourIndex === 0) {
+        const fundingAmount = getFundingAmount(state, state.ourIndex);
+        return {
+          channelState: {
+            ...state,
+            fundingState: fundingStates.aWaitForDepositToBeSentToMetaMask({
+              ...state.fundingState,
+            }),
+          },
+          outboxState: {
+            transactionOutbox: createDepositTransaction(state.channelId, fundingAmount),
+          },
+        };
+      } else {
+        const updatedState = states.waitForFundingConfirmation(state);
+        if (unhandledAction) {
+          return fundingReducer({ ...updatedState }, unhandledAction);
+        } else {
+          return { channelState: updatedState };
+        }
+      }
+    case actions.FUNDING_REJECTED:
+      const sendFundingDeclinedAction = messageRelayRequested(
+        state.participants[1 - state.ourIndex],
+        'FundingDeclined',
+      );
+      return {
+        channelState: states.sendFundingDeclinedMessage({ ...state }),
+        outboxState: { messageOutbox: sendFundingDeclinedAction, displayOutbox: hideWallet() },
+      };
+    case actions.MESSAGE_RECEIVED:
+      if (action.data && action.data === 'FundingDeclined') {
+        return { channelState: states.acknowledgeFundingDeclined(state) };
+      } else {
+        return { channelState: state };
+      }
+    case actions.FUNDING_DECLINED_ACKNOWLEDGED:
+      return { channelState: states.approveFunding({ ...state, unhandledAction: action }) };
+    default:
+      return { channelState: state };
+  }
+};
+
+const aWaitForPostFundSetupReducer = (
+  state: states.AWaitForPostFundSetup,
+  action: actions.WalletAction,
+): NextChannelState<states.ChannelState> => {
+  switch (action.type) {
+    case actions.COMMITMENT_RECEIVED:
+      const { commitment: postFundState, signature } = action;
+      if (!validTransitionToPostFundState(state, postFundState, signature)) {
+        return { channelState: state };
+      }
+
+      return {
+        channelState: states.acknowledgeFundingSuccess({
+          ...state,
+          turnNum: postFundState.turnNum,
+          lastCommitment: { commitment: postFundState, signature },
+          penultimateCommitment: state.lastCommitment,
+        }),
+      };
+    default:
+      return { channelState: state };
+  }
+};
+
+const waitForFundingConfirmationReducer = (
+  state: states.WaitForFundingConfirmation,
+  action: actions.WalletAction,
+): NextChannelState<states.ChannelState> => {
+  switch (action.type) {
+    case actions.FUNDING_RECEIVED_EVENT:
+      const { allocation } = state.lastCommitment.commitment;
+      if (bigNumberify(action.totalForDestination).gte(allocation[1 - state.ourIndex])) {
+        return {
+          channelState: {
+            ...state,
+            fundingState: fundingStates.bWaitForDepositToBeSentToMetaMask({
+              ...state.fundingState,
+            }),
+          },
+          outboxState: {
+            transactionOutbox: createDepositTransaction(
+              state.channelId,
+              allocation[state.ourIndex],
+            ),
+          },
+        };
+      } else {
+        return { channelState: state };
+      }
+    default:
+      return { channelState: state };
+  }
+};
+
+const bWaitForPostFundSetupReducer = (
+  state: states.BWaitForPostFundSetup,
+  action: actions.WalletAction,
+): NextChannelState<states.ChannelState> => {
+  switch (action.type) {
+    case actions.COMMITMENT_RECEIVED:
+      const { commitment, signature } = action;
+      if (!validTransitionToPostFundState(state, commitment, signature)) {
+        return { channelState: state };
+      }
+
+      const newState = { ...state, turnNum: commitment.turnNum };
+      const {
+        postFundSetupCommitment,
+        commitmentSignature,
+        sendCommitmentAction,
+      } = composePostFundCommitment(newState);
+      return {
+        channelState: states.acknowledgeFundingSuccess({
+          ...newState,
+          turnNum: postFundSetupCommitment.turnNum,
+          lastCommitment: { commitment: postFundSetupCommitment, signature: commitmentSignature },
+          penultimateCommitment: { commitment, signature },
+        }),
+        outboxState: { messageOutbox: sendCommitmentAction },
+      };
+    default:
+      return { channelState: state };
+  }
+};
+
+const acknowledgeFundingSuccessReducer = (
+  state: states.AcknowledgeFundingSuccess,
+  action: actions.WalletAction,
+): NextChannelState<states.ChannelState> => {
+  switch (action.type) {
+    case actions.FUNDING_SUCCESS_ACKNOWLEDGED:
+      return {
+        channelState: states.waitForUpdate({
+          ...state,
+        }),
+        outboxState: {
+          displayOutbox: hideWallet(),
+          messageOutbox: fundingSuccess(state.channelId, state.lastCommitment.commitment),
+        },
+      };
+    default:
+      return { channelState: state };
+  }
+};
+
+const validTransitionToPostFundState = (
+  state: states.FundingChannelState,
+  data: Commitment,
+  signature: string | undefined,
+) => {
+  if (!signature) {
+    return false;
+  }
+
+  const opponentAddress = state.participants[1 - state.ourIndex];
+
+  if (!validCommitmentSignature(data, signature, opponentAddress)) {
+    return false;
+  }
+  if (!validTransition(state, data)) {
+    return false;
+  }
+  if (data.commitmentType !== 1) {
+    return false;
+  }
+  return true;
+};
+
+const composePostFundCommitment = (
+  state: states.WaitForFundingConfirmation | states.BWaitForPostFundSetup,
+) => {
+  const { libraryAddress, channelNonce, participants, turnNum, lastCommitment } = state;
+  const channel: Channel = { channelType: libraryAddress, nonce: channelNonce, participants };
+
+  const postFundSetupCommitment: Commitment = {
+    channel,
+    commitmentType: CommitmentType.PostFundSetup,
+    turnNum: turnNum + 1,
+    commitmentCount: state.ourIndex,
+    allocation: lastCommitment.commitment.allocation,
+    destination: lastCommitment.commitment.destination,
+    appAttributes: state.lastCommitment.commitment.appAttributes,
+  };
+  const commitmentSignature = signCommitment(postFundSetupCommitment, state.privateKey);
+
+  const sendCommitmentAction = commitmentRelayRequested(
+    state.participants[1 - state.ourIndex],
+    postFundSetupCommitment,
+    commitmentSignature,
+  );
+  return { postFundSetupCommitment, commitmentSignature, sendCommitmentAction };
+};
+
+const getFundingAmount = (state: states.FundingChannelState, index: number): string => {
+  const lastCommitment = state.lastCommitment.commitment;
+  return lastCommitment.allocation[index];
+};
